Instead of a config.php, now using a .env for better separation between environments
Also started to implemented a properly helpers file for better global handling without including the file always
Here is work to done, the most annoying part is to remove the PageController includes, but thats results into a complet code reformat

I added a git ignore file for removing sensitive files or other files like the vendor directory.

I also had to set the key `session-token` in the page controller to zero if it is not present in the cookies.